Add test coverage for our package exports
