+++ .babelrc
@@ -3,6 +3,7 @@
     "plugins": [
         "transform-flow-strip-types",
         "transform-class-properties",
-        "transform-object-rest-spread"
+        "transform-object-rest-spread",
+        "transform-dynamic-import"
     ]
 }
