+++ packages/wonder-blocks-button/components/button.test.js
@@ -1,0 +1,14 @@
+// @flow
+import React from "react";
+import {shallow} from "enzyme";
+
+import Button from "./button.js";
+
+describe("Button", () => {
+    test("render a label", (done) => {
+        const wrapper = shallow(
+            <Button onClick={() => done()}>Hello World!</Button>,
+        );
+        wrapper.simulate("click");
+    });
+});
