+++ packages/wonder-blocks-color/index.test.js
@@ -1,0 +1,18 @@
+// @flow
+describe("@khanacademy/wonder-blocks-color", () => {
+    test("package exports default, SemanticColor, mix, and fade", async () => {
+        // Arrange
+        const importedModule = import("./index.js");
+
+        // Act
+        const result = await importedModule;
+
+        // Assert
+        expect(Object.keys(result)).toEqual([
+            "default",
+            "SemanticColor",
+            "mix",
+            "fade",
+        ]);
+    });
+});
